The carpet still isn't working, but I added all of the things that I think I need for it, and commented out the line where the carpet is added to the environment.

+++ SkyBox.h
@@ -165,7 +165,7 @@ public:
 	}
 
 	float get_max_brick_coord(float brick_radius) {
-		return scale * 0.5 - brick_radius;
+		return (float)(scale * 0.5f - brick_radius);
 	}
 };
 #endif
